Move CodeTheme to DocumentationTopic
